+++ NEWS.md
@@ -1,6 +1,7 @@
 ## Changes in version 0.2.1
 
 * Minor CRAN request from 2022-02-17, concerning the package logo; fixed.
+* Minor change, make_Sample, class() query corrected
 
 ## Changes in version 0.2.0